Add Lodash to group by focus area
This code adds lodash and groups the CSV data by Focus Area. It displays an h1 for each focus area, then lists some data for each grant.

The point is just to display the data we're working with; as such, it's probably over-engineered, but no big deal.

Perhaps the next step to take is to create an SVG with a pie chart for each focus area, with each grant's data going as a slice of its respective focus area's pie.